Development

+++ bot/database/models/channels.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Boolean
+from sqlalchemy import BigInteger, String, Boolean
 from sqlalchemy.orm import Mapped, mapped_column
 from bot.database.base import Base
 
@@ -6,8 +6,8 @@ from bot.database.base import Base
 class Channel(Base):
     __tablename__ = "channels"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    tg_id: Mapped[int] = mapped_column(Integer)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    tg_id: Mapped[int] = mapped_column(BigInteger)
     channel_link: Mapped[str] = mapped_column(String, nullable=True)
     channel_username: Mapped[str] = mapped_column(String, nullable=True)
     channel_title: Mapped[str] = mapped_column(String, nullable=False)
